Made GoogleMapGateway a facade in order to adhere to SRP.
